fix: bug fix on toString() method

+++ labs/comp2522 lab 3/src/code/ca/bcit/comp2522/lab3/device/IPod.java
@@ -6,12 +6,14 @@ import java.util.Objects;
  * Represents an iPod.
  *
  * @author Aika Manalo - 2C
+ * @author Julia Ziebart - 2C
+ *
  * @version 1.0
  */
 public class IPod extends IDevice
 {
-    private static final int MAX_VOLUME_DB = 100;
-    private static final int MIN_VOLUME_DB = 0;
+    private static final double MAX_VOLUME_DB = 100.0;
+    private static final double MIN_VOLUME_DB = 0.0;
     private static final int MIN_NUM_SONGS = 1;
 
     private final double maxVolumeDb;
